Reorganized source files to fit uv's pkg model
