+++ pyproject.toml
@@ -16,3 +16,10 @@ dependencies = [
 dev = [
     "black>=25.1.0",
 ]
+
+[project.scripts]
+furrit = "furrit:main"
+
+[build-system]
+requires = ["uv_build>=0.8.13,<0.9.0"]
+build-backend = "uv_build"
